% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filterDate}
\alias{filterType}
\title{Filter the Data with the Bond Type and Date}
\usage{
filterType(data, type)

filterDate(data, start, end, date_type = "initdate")
}
\arguments{
\item{data}{data.frame, it's better to be returned by \code{getDataUsingWind}
however if the data has the column *type* it will work}

\item{type}{character, used to choose which type of credit bond to be
filtered, if the type is incorrect, the function will stop and give message}

\item{start}{the start date, can be string like "2016-01-01"}

\item{end}{the end date, can be string like "2016-01-01"}

\item{date_type}{the date type used to filtered data, "initdate" or
"carrydate"}
}
\value{
The filtered data.frame for the further work.
}
\description{
Filter the data with the attribute of bond's type and important dates, this
will help to
customize the further work. It is the best practice to use this filter
function to the data returned by \code{getDataUsingWind}.
}
\details{
Pay attention to the point that
}

