% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataUsingWind.R
\name{getDataUsingWind}
\alias{getDataUsingWind}
\title{Read Data From Wind Client}
\usage{
getDataUsingWind(file)
}
\arguments{
\item{file}{the file name}
}
\value{
a data.frame
}
\description{
Read Data From Wind Client
}
\details{
Read Data From Wind Client
}

