% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{This is the Title.}
\usage{
hello()
}
\arguments{
\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}
}
\description{
This is the description.
}
\details{
These are further details. What's going on here?
Text accompanying the custom section.
}

