% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_im.R
\name{kable_im}
\alias{kable_im}
\title{The Improvement of \code{knitr::kable} Function}
\usage{
kable_im(data, caption, split = TRUE, ignore = 1, align = "c",
  align.str = "", font.size = "")
}
\arguments{
\item{data}{data.frame, the table wait to be converted to Latex code}

\item{caption}{character, the caption of the table}

\item{split}{logical, whether to split the header to tow lines}

\item{ignore}{numeric, which column(s) to be ignored when splitting}

\item{align}{character, how to align the columns, "c", "r" or "custom"}

\item{align.str}{if \code{align} is set to "custom", the customized the
alignment should be given here}

\item{font.size}{the Latex code of font size can be set directly here, e.g.
"\\small"}
}
\value{
The LaTex code of the table will be output using \code{cat}, thus it
  can be copied to Latex file or used in **knitr**, where set
  \code{result='asis'}, the same way as using \code{kable}.
}
\description{
The improved and customized version of the `kable` function which is used to
generated Latex table. Generally speaking, it add more customization options.
}
\details{
The function uses \code{kable} to generate table in the **table** environment,
but lack of customization of the table style. This funcion adds the following
options:

- The font size can be directly input to the final Latex code.

- Set or give the alignment option of the table columns. Like "l|c|c|r".

- Split the header to tow lines, which will solve the wide table to be
located inside the text area.
}

