% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRank.R
\name{printRank}
\alias{printRank}
\title{Print The Ranking Result}
\usage{
printRank(data, header_cn = TRUE)
}
\arguments{
\item{data}{the input ranking data, returned by \code{uwRanking}}

\item{header_cn}{logical, if TRUE, default, the returned data.frame will have
a Chinese header}
}
\value{
A prettified data.frame which is exactly the same as input data,
  however friendlier to be viewed.
}
\description{
Prettify and print the ranking result
}
\details{
This function prettify the ranking result returned by \code{uwRanking}. First
of all, the data structure will be checked, if the warning message occurs,
make sure the data is returned by \code{uwRanking} function. Then, the
numeric columns are cut to 2 digits to the right of decimal. And the
percentage figure is multiplied 100 and pasted "%". At last, the underwriter
names are justified with full space, i.e., if the longest underwriter name is
5 chars, then the names with 4 chars will be added one full space to the
right side. The column names can be set to Chinese with the parameter.
}

