% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pryrk.R
\docType{package}
\name{pryrk}
\alias{pryrk}
\alias{pryrk-package}
\title{pryrk: Primary Bond Market Ranking in China}
\description{
This package is designed specifically for the people who concern
the bond market of China. The functions are used to obtain the ranking of
underwriters' performance in the primary bond market. However, the data used
should be get from the Wind client. And, Wind is data provider in China,
only the user of Wind can have the access to their data.
}

