% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{rank}
\alias{rank}
\title{Quick Rank Function}
\usage{
rank(envir = .GlobalEnv, file, update = FALSE, save = FALSE, type, start,
  end, date_type, output, sheet = "Rank", open = FALSE, print.head = TRUE)
}
\arguments{
\item{envir}{the name of the environment save and get the data, default is
the Global Environment}

\item{file}{the file name of Excel data file}

\item{update}{logical, default is FALSE, if TRUE the function will be forced
to read the Excel file to get the original bond data, otherwise, it will
check if there exists the data and use it without reading again}

\item{save}{logical, default FALSE, when TRUE the rank result will be saved
to Excel file}

\item{type}{character, used to choose which type of credit bond to be
filtered, if the type is incorrect, the function will stop and give message}

\item{start}{the start date, can be string like "2016-01-01"}

\item{end}{the end date, can be string like "2016-01-01"}

\item{date_type}{the date type used to filtered data, "initdate" or
"carrydate"}

\item{output}{the file name of Excel, used when param **save** is TRUE}

\item{sheet}{the sheet name, default "Rank"}

\item{open}{logical, if TRUE, the output file will be opened, default is
FALSE}

\item{print.head}{logical, default TRUE, when the function is called, the
top 20 will be printed as referenced}
}
\value{
The ranking result, can be saved into an Excel file.
}
\description{
Quickly obtain the ranking result with many params have been given usually
used values. Of course, it depends on the original Excel data file.
}
\details{
This function comprehensively use other functions in the package and quickly
gives the underwriter result. Though, there are many params, most of which
are pass to the functions used, and in fact, they have the default value.
That is to say that if you just type \code{rank()}, it will read and
compute the ranking result. At this time, the result is fully based on the
input data file, however you are sure to allow add filter params later.

The function will show the top 20 by default, and surely can save the result
to Excel file, at the same time **assign** the result to the *envir* given,
which can be found as named \code{uwrk} ("Underwriters Ranking").
}

