% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankNumber.R
\name{rankNumber}
\alias{rankNumber}
\title{Give the Rank Number}
\usage{
rankNumber(vector, decreasing = TRUE)
}
\arguments{
\item{vector}{numeric, the vector will be sorted and given the rank number}

\item{decreasing}{logical, default is TRUE, used in function \code{sort}}
}
\value{
The rank vector of input.
}
\description{
Give the rank number to input numeric vector, the same rank number to the
same value.
}
\details{
For the same value in an ordered vector, the rank number should be the same,
the next value will be given its right ranking number, like, 1, 2, 3, 3, 5...
}

