% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToExcel.R
\name{saveToExcel}
\alias{saveToExcel}
\title{Save the Ranking Result to Excel File}
\usage{
saveToExcel(data, file, sheet = "Rank", open = FALSE)
}
\arguments{
\item{data}{the ranking data returned by \code{uwRanking}}

\item{file}{the output file name, default "primaryRanking.xlsx"}

\item{sheet}{the sheet name, default "Rank"}

\item{open}{logical, if TRUE, the output file will be opened, default is
FALSE}
}
\value{
The input data will be saved to an Excel file.
}
\description{
Save the ranking result returned by \code{uwRanking} to the Excel file. This
function just save the single ranking dara.frame to single worksheet. The
cell style will be set
}

