% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwRanking.R
\name{uwRanking}
\alias{uwRanking}
\title{Compute the Rank of Underwriter}
\usage{
uwRanking(data)
}
\arguments{
\item{data}{the input data which is returned by \code{uwSplit}}

\item{header_cn}{logical, if \code{TRUE} then the result's column names will
be in Chinese, default is \code{FALSE}. Now suspended}

\item{round}{logical, whether to round the figure to 2 decimal, which will
make it easier to view the results, however the precision is not so
important for this work, default is TRUE. Now suspended}
}
\value{
The data.frame containing the ranking result.
}
\description{
Using the data processed by \code{uwSplit} to compute the ranking result of
underwriters.
}
\details{
The result includes three indexes, weighted-mean of quantity and volume, and
the absolute quantity of underwriting performance. The result use the
weighted mean as the order, however it includes all the ranking number,
figure and ratio.
}

