% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwSplit.R
\name{uwSplit}
\alias{uwSplit}
\title{Process the Data for Rank}
\usage{
uwSplit(data, use.tidyr = TRUE)
}
\arguments{
\item{data}{the original primary bond market data, which is read by the
function \code{getDataUsingWind}, and it must contain column "underwriter."}

\item{use.tidyr}{logical, whether to use \code{tidyr::separate_rows}}
}
\value{
A data.frame like the input one, however, every record has only one
  underwriter. And this data.frame will be used for ranking.
}
\description{
Process the original primary bond market data for the rank, especially split
the co-underwriting bond record into several records according to the number
of the underwriter.
}
\details{
The data returned by \code{getDataUsingWind} is used here. This is the step
before ranking. It is easy to handle with the single underwriter record,
however the co-underwriters are "pasted" into on string. This function split
co-underwriters as they are seperated by ",". The param \code{use.tidyr} is
TRUE in default, and this will use \code{tidyr::separate_rows} which will
improve the speed obviously.
}

