% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filterDate}
\alias{filterType}
\title{Filter the Data with the Bond Type and Date}
\usage{
filterType(data, type)

filterDate(data, start = NULL, end = NULL, date_type = "initdate")
}
\arguments{
\item{data}{data.frame, it's better to be returned by \code{getDataUsingWind}
however if the data has the column *type* it will work}

\item{type}{character, used to choose which type of credit bond to be
filtered, if the type is incorrect, the function will stop and give message}

\item{start}{the start date, can be string like "2016-01-01"}

\item{end}{the end date, can be string like "2016-01-01"}

\item{date_type}{the date type used to filtered data, "initdate" or
"carrydate"}
}
\value{
The filtered data.frame for the further work.
}
\description{
Filter the data with the attribute of bond's type and important dates, this
will help to
customize the further work. It is the best practice to use this filter
function to the data returned by \code{getDataUsingWind}. But if the data
returned by \code{uwSplit} exists and such data is big, then use filter to it
is better to avoid the long-time wait by using \code{uwSplit}.
}
\details{
Pay attention to that when filtering the type, only three type options can
be given and get the right result. When filtering date, start and end
dates are all given and obtain the result, otherwise two params are default
set to \code{NULL} which will just pass throught the input \code{data}.
}

