% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataUsingWind.R
\name{getDataUsingWind}
\alias{getDataUsingWind}
\title{Read Data From Wind Client}
\usage{
getDataUsingWind(file)
}
\arguments{
\item{file}{A "character" string giving the file name path.}
}
\value{
A "clean" data.frame which can be used for further use.
}
\description{
Read primary bond market data from Wind client and return the processed data.
}
\details{
This functions is used to read the Excel file which is exported by Wind
client. Only the columns useful for the further work will be maintain in the
result \code{data.frame}. Reading Excel file uses the package \code{readxl}.
}

